+++ forui/lib/src/widgets/toast/toaster_style.dart
@@ -39,25 +39,17 @@ class FToasterStyle with Diagnosticable, _$FToasterStyleFunctions {
   @override
   final Duration expandHoverEnterDuration;
 
-  /// The duration to wait after exiting the toaster before collapsing the toasts. Defaults to 300ms.
+  /// The duration to wait after exiting the toaster before collapsing the toasts. Defaults to 200ms.
   @override
   final Duration expandHoverExitDuration;
 
-  /// The spacing below or above the toasts when they are expanded. Defaults to 16.0.
+  /// The spacing below or above the toasts when they are expanded. Defaults to 0.
   @override
   final double expandStartSpacing;
 
   /// The spacing between the toasts when they are expanded. Defaults to 10.0.
   @override
   final double expandSpacing;
-
-  /// The expanding/collapsing animation duration. Defaults to 500ms.
-  @override
-  final Duration expandDuration;
-
-  /// The expanding/collapsing animation curve. Defaults to [Curves.easeInOutCubic].
-  @override
-  final Curve expandCurve;
 
   /// The protrusion of the collapsed toasts behind the front toast. This is scaled by the number of toasts in
   /// front of the toast.
@@ -72,6 +64,10 @@ class FToasterStyle with Diagnosticable, _$FToasterStyleFunctions {
   /// Defaults to 0.9.
   @override
   final double collapsedScale;
+
+  /// The motion-related properties.
+  @override
+  final FToasterMotion motion;
 
   /// The toast's alignment relative to a [FToaster]. Defaults to [FToastAlignment.bottomEnd].
   @override
@@ -88,13 +84,12 @@ class FToasterStyle with Diagnosticable, _$FToasterStyleFunctions {
     this.padding = const EdgeInsets.symmetric(horizontal: 20, vertical: 15),
     this.expandBehavior = FToasterExpandBehavior.hoverOrPress,
     this.expandHoverEnterDuration = const Duration(milliseconds: 200),
-    this.expandHoverExitDuration = const Duration(milliseconds: 300),
-    this.expandStartSpacing = 16,
+    this.expandHoverExitDuration = const Duration(milliseconds: 200),
+    this.expandStartSpacing = 0,
     this.expandSpacing = 10,
-    this.expandDuration = const Duration(milliseconds: 500),
-    this.expandCurve = Curves.easeInOutCubic,
     this.collapsedProtrusion = 12,
     this.collapsedScale = 0.9,
+    this.motion = const FToasterMotion(),
     this.toastAlignment = FToastAlignment.bottomEnd,
   });
 
@@ -105,45 +100,35 @@ class FToasterStyle with Diagnosticable, _$FToasterStyleFunctions {
       );
 }
 
+/// The motion-related properties for [FToaster] that affect all toasts.
+class FToasterMotion with Diagnosticable, _$FToasterMotionFunctions {
+  /// The duration of the toasts' expansion. Defaults to 400ms.
+  @override
+  final Duration expandDuration;
+
+  /// The duration of the toasts' collapsing. Defaults to 300ms.
+  @override
+  final Duration collapseDuration;
+
+  /// The animation curve for the toasts' expansion and collapsing. Defaults to [Curves.easeOutCubic].
+  @override
+  final Curve expandCurve;
+
+  /// The animation curve for the toasts' collapsing. Defaults to [Curves.easeOutCubic].
+  @override
+  final Curve collapseCurve;
+
+  /// Creates a [FToasterMotion].
+  const FToasterMotion({
+    this.expandDuration = const Duration(milliseconds: 400),
+    this.collapseDuration = const Duration(milliseconds: 300),
+    this.expandCurve = Curves.easeOutCubic,
+    this.collapseCurve = Curves.easeOutCubic,
+  });
+}
+
 /// The toast's style.
 class FToastStyle with Diagnosticable, _$FToastStyleFunctions {
-  /// The toast's entrance & exit animation duration. Defaults to 400ms.
-  @override
-  final Duration enterExitDuration;
-
-  /// The toast's entrance animation curve. Defaults to [Curves.easeOutCubic].
-  @override
-  final Curve enterCurve;
-
-  /// The toast's exit animation curve. Defaults to [Curves.easeOutCubic].
-  @override
-  final Curve exitCurve;
-
-  /// The toast's initial opacity when it enters, and the target opacity when it exits.
-  ///
-  /// Defaults to 0. Set to 1.0 to remove the fade-in/out effect.
-  ///
-  /// ## Contract
-  /// Throws [AssertionError] if the value is not in `[0, 1]`.
-  @override
-  final double entranceExitOpacity;
-
-  /// The toast's transition between indexes animation duration. Defaults to 400ms.
-  @override
-  final Duration transitionDuration;
-
-  /// The toast's transition animation curve. Defaults to [Curves.easeInOutCubic].
-  @override
-  final Curve transitionCurve;
-
-  /// The toast's swipe completion animation duration. Defaults to 150ms.
-  @override
-  final Duration swipeCompletionDuration;
-
-  /// The toast's swipe completion animation curve. Defaults to [Curves.easeInCubic].
-  @override
-  final Curve swipeCompletionCurve;
-
   /// The toast's constraints. Defaults to `BoxConstraints(maxHeight: 250, maxWidth: 400)`.
   @override
   final BoxConstraints constraints;
@@ -208,26 +193,23 @@ class FToastStyle with Diagnosticable, _$FToastStyleFunctions {
   @override
   final double suffixSpacing;
 
+  /// The motion-related properties.
+  @override
+  final FToastMotion motion;
+
   /// Creates a [FToastStyle].
   FToastStyle({
     required this.decoration,
     required this.iconStyle,
     required this.titleTextStyle,
     required this.descriptionTextStyle,
-    this.enterExitDuration = const Duration(milliseconds: 400),
-    this.enterCurve = Curves.easeOutCubic,
-    this.exitCurve = Curves.easeOutCubic,
-    this.entranceExitOpacity = 0.0,
-    this.transitionDuration = const Duration(milliseconds: 400),
-    this.transitionCurve = Curves.easeOutCubic,
-    this.swipeCompletionDuration = const Duration(milliseconds: 150),
-    this.swipeCompletionCurve = Curves.easeInCubic,
     this.constraints = const BoxConstraints(maxHeight: 250, maxWidth: 400),
     this.padding = const EdgeInsets.all(16),
     this.backgroundFilter,
     this.iconSpacing = 10,
     this.titleSpacing = 1,
     this.suffixSpacing = 12,
+    this.motion = const FToastMotion(),
   });
 
   /// Creates a [FToastStyle] that inherits its properties.
@@ -243,4 +225,82 @@ class FToastStyle with Diagnosticable, _$FToastStyleFunctions {
         titleSpacing: 5,
         descriptionTextStyle: typography.sm.copyWith(color: colors.mutedForeground, overflow: TextOverflow.ellipsis),
       );
+}
+
+/// The motion-related properties for [FToaster] that affect individual toasts.
+class FToastMotion with Diagnosticable, _$FToastMotionFunctions {
+  /// The duration of the toast's entrance when it is initially added to to toaster. Defaults to 400ms.
+  @override
+  final Duration entranceDuration;
+
+  /// The duration of the toast's exit animation when it is dismissed. Defaults to 300ms.
+  @override
+  final Duration dismissDuration;
+
+  /// The duration of the toast's transition between places in the toaster. Defaults to 400ms.
+  @override
+  final Duration transitionDuration;
+
+  /// The duration of the toast's fade-in animation when another toast has been dismissed and this toast re-enters the
+  /// toaster. Defaults to 400ms.
+  @override
+  final Duration reentranceDuration;
+
+  /// The duration of the toast's fade-out animation when the number of toasts in a toaster exceeds the maximum allowed
+  /// and this toast is hidden. Defaults to 400ms.
+  @override
+  final Duration exitDuration;
+
+  /// The toast's swipe completion animation duration. Defaults to 150ms.
+  @override
+  final Duration swipeCompletionDuration;
+
+  /// The toast's initial entrance animation curve. Defaults to [Curves.easeOutCubic].
+  @override
+  final Curve entranceCurve;
+
+  /// The toast's exit animation curve. Defaults to [Curves.easeOutCubic].
+  @override
+  final Curve dismissCurve;
+
+  /// The toast's transition animation curve. Defaults to [Curves.easeOutCubic].
+  @override
+  final Curve transitionCurve;
+
+  /// The curve of the toast's fade-in animation when another toast has been dismissed and this toast re-enters the
+  /// toaster. Defaults to [Curves.easeOutCubic].
+  @override
+  final Curve reentranceCurve;
+
+  /// The curve of the toast's fade-out animation when the number of toasts in a toaster exceeds the maximum allowed
+  /// and this toast is hidden. Defaults to [Curves.easeOutCubic].
+  @override
+  final Curve exitCurve;
+
+  /// The toast's swipe completion animation curve. Defaults to [Curves.easeInCubic].
+  @override
+  final Curve swipeCompletionCurve;
+
+  /// The toast's initial entrance's opacity and dismiss's fade tween. Defaults to `[0, 1]`.
+  ///
+  /// Set to `[1, 1]` to disable the fade-in/out effect.
+  @override
+  final Animatable<double> entranceDismissFadeTween;
+
+  /// Creates a [FToastMotion].
+  const FToastMotion({
+    this.entranceDuration = const Duration(milliseconds: 400),
+    this.dismissDuration = const Duration(milliseconds: 300),
+    this.transitionDuration = const Duration(milliseconds: 400),
+    this.reentranceDuration = const Duration(milliseconds: 400),
+    this.exitDuration = const Duration(milliseconds: 400),
+    this.swipeCompletionDuration = const Duration(milliseconds: 150),
+    this.entranceCurve = Curves.easeOutCubic,
+    this.dismissCurve = Curves.easeOutCubic,
+    this.transitionCurve = Curves.easeOutCubic,
+    this.reentranceCurve = Curves.easeOutCubic,
+    this.exitCurve = Curves.easeOutCubic,
+    this.swipeCompletionCurve = Curves.easeInCubic,
+    this.entranceDismissFadeTween = const FImmutableTween(begin: 0.0, end: 1.0),
+  });
 }